refactor(Common): Separate OptionPreferences into its own files in Core
