+++ Core/GameEngine/CMakeLists.txt
@@ -82,6 +82,7 @@ set(GAMEENGINE_SRC
 #    Include/Common/MultiplayerSettings.h
 #    Include/Common/NameKeyGenerator.h
     Include/Common/ObjectStatusTypes.h
+    Include/Common/OptionPreferences.h
 #    Include/Common/OSDisplay.h
 #    Include/Common/Overridable.h
 #    Include/Common/Override.h
@@ -613,6 +614,7 @@ set(GAMEENGINE_SRC
 #    Source/Common/MiniLog.cpp
 #    Source/Common/MultiplayerSettings.cpp
 #    Source/Common/NameKeyGenerator.cpp
+    Source/Common/OptionPreferences.cpp
 #    Source/Common/PartitionSolver.cpp
 #    Source/Common/PerfTimer.cpp
     Source/Common/RandomValue.cpp
