+++ Generals/Code/GameEngine/Source/Common/Recorder.cpp
@@ -45,7 +45,7 @@
 #include "GameLogic/GameLogic.h"
 #include "Common/RandomValue.h"
 #include "Common/CRCDebug.h"
-#include "Common/UserPreferences.h"
+#include "Common/OptionPreferences.h"
 #include "Common/version.h"
 
 constexpr const char s_genrep[] = "GENREP";
