+++ Generals/Code/GameEngine/Source/GameClient/GUI/GUICallbacks/Menus/SkirmishGameOptionsMenu.cpp
@@ -256,7 +256,7 @@ Int SkirmishPreferences::getPreferredFaction(void)
 
 Bool SkirmishPreferences::usesSystemMapDir(void)
 {
-	OptionPreferences::const_iterator it = find("UseSystemMapDir");
+	SkirmishPreferences::const_iterator it = find("UseSystemMapDir");
 	if (it == end())
 		return TRUE;
